Create ExecutionTime.cpp
The problem you linked is about calculating the exclusive time of functions based on log entries. Here's an approach to solving this problem using a stack:

Create a stack named stk to store the function execution start times.
Create a vector named ans of size n (the number of functions) and initialize all elements to 0. This vector will store the exclusive time for each function.
Iterate through the log entries:
If the log entry indicates the start of a function execution:
If the stack is not empty, update the exclusive time of the function at the top of the stack. Add the difference between the current timestamp and the top element of the stack to ans[stk.top().first].
Push a pair to the stack with the function identifier (logs[0][i]) and the current timestamp (logs[2][i]).
If the log entry indicates the end of a function execution:
Retrieve the start time of the function from the top of the stack (x = stk.top().second).
Calculate the exclusive time for the current function by subtracting x from the current timestamp and adding 1 (t = logs[2][i] - x + 1).
Add the exclusive time to ans[logs[0][i]].
Pop the top element from the stack.
If the stack is not empty, update the start time of the new top element to the current timestamp plus 1 (stk.top().second = logs[2][i] + 1).
Return the ans vector containing the exclusive time for each function.
This approach utilizes a stack to keep track of the start times of function executions. When an end time is encountered, the start time is retrieved from the stack to calculate the exclusive time for the function. By updating the exclusive time for each function during the iteration, we can obtain the desired result.